lazily count the files/folders in a directory.
previous impl already lazily counted the dir (only if visible), however this would cause opening a new dir
to take too long as all 9 new entries are visible and counted. this was made worse by the fact that in the same
frame, the directory would be scanned, which usually takes 2ms+. this would result in the whole frame taking 16-19ms+
which results in a dropped frame.

the new method only counts 1 folder per frame.

a better fix would be to keep track of the time spent / remaining in a frame and use that to know whether it
should attempt to count another folder. ideal if the previous few dirs were empty and thus would take no time at all.